Extract shell published state merger
